final-pjt-back/requirements.txt
asgiref==3.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.12
django-allauth==0.50.0
django-appconf==1.0.5
django-cors-headers==3.12.0
django-environ==0.8.1
django-extensions==3.1.5
django-imagekit==4.1.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
idna==3.3
oauthlib==3.2.0
pilkit==2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
